% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategicplayers.R
\name{distance}
\alias{distance}
\title{distance}
\usage{
distance(gd, targets, avoiders, theta, players)
}
\arguments{
\item{gd}{a matrix of geodesic distances for the network of interest}

\item{targets}{a vector of indicies of the people you want to spread the intervention to}

\item{avoiders}{a vector of indicies of the people you don't want to spread the intervention to}

\item{theta}{a number between 0 and 1 which weights the distance metric, 1 only prioritizes closeness to targets, 0 only prioritizes maximizing distance from avoiders}

\item{players}{the indicies of people who you have chosen for the intervention (a subset of targets)}
}
\value{
returns the distance metric for strategic players, which we want to maximize
}
\description{
Takes in the geodesic distances, targets, avoiders, a parameter that prioritizes avoiding vs targetting, and the current players and returns the strategic players distance metric
}

